hid-steam: Add basic support for Steam Deck controller
This commit extends the Steam controller driver with support for the
Steam Deck. The Steam Deck controller expects to get 64 byte long
messages, so every buffer is extended to that size. This fixes the
lizard mode disabling. On top of that the Steam Deck controller always
reverts back to lizard mode after a few seconds. I traced that Steam
sends periodic disabling messages so I have created a crude kernel
thread once the lizard mode is deactivated to repeatedly disable it.

Additionally this commit adds new mapping for the Steam Deck Controller
which is a bit different from the old controller.

Signed-off-by: Hendrik Borghorst <root@hendrik-steamdeck.djselbeck.de>